fix: behave sensibly when there are 0 pages / no results
